+++ iOS/LeafByte/Images/ImageConversions.swift
@@ -23,40 +23,45 @@ func cgToUiImage(_ cgImage: CGImage) -> UIImage {
 
 // Convert a Core Image image to a Core Graphics image.
 let context: CIContext = CIContext(options: nil)
-func ciToCgImage(_ ciImage: CIImage) -> CGImage {
+func ciToCgImage(_ ciImage: CIImage) -> CGImage? {
     // It's appealing to lead with something like:
     // if #available(iOS 10.0, *), ciImage.cgImage != nil {
     //     return ciImage.cgImage!
     // }
     // But it turns out that this gives you a downsampled version.
     
-    return context.createCGImage(ciImage, from: ciImage.extent)!
+    return context.createCGImage(ciImage, from: ciImage.extent)
 }
 
 // Convert a Core Image image to a UI image.
-func ciToUiImage(_ ciImage: CIImage) -> UIImage {
+func ciToUiImage(_ ciImage: CIImage) -> UIImage? {
     // UIImage(ciImage: ciImage) seems obvious, but it stretches the image.
     
-    let cgImage = ciToCgImage(ciImage)
+    guard let cgImage = ciToCgImage(ciImage) else {
+        return nil
+    }
     return UIImage(cgImage: cgImage)
 }
 
 // Convert a UI image to a Core Graphics image.
-func uiToCgImage(_ uiImage: UIImage) -> CGImage {
+func uiToCgImage(_ uiImage: UIImage) -> CGImage? {
     // It's appealing to lead with something like:
     // if uiImage.cgImage != nil {
     //     return uiImage.cgImage!
     // }
     // But it turns out that this gives you a downsampled version.
     
-    return ciToCgImage(uiToCiImage(uiImage))
+    guard let ciImage = uiToCiImage(uiImage) else {
+        return nil
+    }
+    return ciToCgImage(ciImage)
 }
 
 // Convert a UI image to a Core Image image.
-func uiToCiImage(_ uiImage: UIImage) -> CIImage {
+func uiToCiImage(_ uiImage: UIImage) -> CIImage? {
     if uiImage.ciImage != nil {
        return uiImage.ciImage!
     }
     
-    return CIImage(image: uiImage)!
+    return CIImage(image: uiImage)
 }